+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## master
 
 * [bugfix] Catch Conflict 409 when creating or updating endpoints
+* [improvement] shelly info should print cloud region
 * [improvement] shelly add should ask about organization name instead of number
   from the list
 * [improvement] Replaced zone with region